Added weighting matrix, restructured code.
